example_clients/opcua_zerobus/advanced/requirements.txt
asyncua>=1.1.0
databricks-zerobus-ingest-sdk>=1.1.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
pika>=1.3.0
